cerbero/meson: Install ninja and meson as build-tools
ninja is not built; a binary is unpacked and installed. There is
a grotesque hack in utils.shell.unpack because ninja.exe does not come
inside a directory, so it extracts to ~/cerbero/sources/build-tools
instead of ~/cerbero/sources/build-tools/ninja-1.6.0. This needs to be
fixed later.

meson is not installed to ~/cerbero/build-tools because setup.py refused
to install there because it contains a check that is broken in our mingw
setup. Instead, we install it in the Python 3 installation prefix. This
is fine overally because doing that with `pip` was part of the original
instructions.

Due to all these changes, ninja now only works when targetting Windows
and Meson only works when running on Windows.

Also moved the checks for Meson and Ninja into configure() because the
recipe-level functions are called in the environment of in which
cerbero-uninstalled is run. Which doesn't have ~/cerbero/build-tools/bin
in PATH.